+++ src/components/Icon/canonicalIconNames.tsx
@@ -1,14 +1,6 @@
 import * as icons from "@carbon/react/icons";
 import { CarbonIconType as IconType } from "@carbon/react/icons";
 
-export type CarbonIconType = IconType;
-
-/**
- * @deprecated
- * use CarbonIconType instead
- */
-export type IconSized = CarbonIconType;
-
 const canonicalIcons = {
     "application-activities": icons.Activity,
     "application-dataintegration": icons.DataUnstructured,
@@ -180,6 +172,8 @@ const canonicalIcons = {
     undefined: icons.Undefined,
 };
 
-export type ValidIconName = keyof typeof canonicalIcons;
 const canonicalIconNames: Record<ValidIconName, CarbonIconType> = canonicalIcons;
+
+export type CarbonIconType = IconType;
+export type ValidIconName = keyof typeof canonicalIcons;
 export default canonicalIconNames;